% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRingFromImage.R
\name{addRingFromImage}
\alias{addRingFromImage}
\title{Add ring to pith to bark profile from CT scan image}
\usage{
addRingFromImage(n = 1, densProfile, im)
}
\arguments{
\item{n}{Number of rings to add}

\item{densProfile}{Density profile}

\item{im}{Density matrix}
}
\value{
Corrected density profile with new ring(s) added and blue bar in plot of added ring
}
\description{
Add ring to pith to bark profile from CT scan image
}
