% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRingToProfile.R
\name{addRingFromProfile}
\alias{addRingFromProfile}
\title{Add ring to pith to bark profile from profile plot}
\usage{
addRingFromProfile(n = 1, densProfile)
}
\arguments{
\item{n}{Number of rings to add}

\item{densProfile}{Density profile}
}
\value{
Corrected density profile with new ring(s) added and blue bar in plot of added ring
}
\description{
Add ring to pith to bark profile from profile plot
}
