% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addYears.R
\name{addYears}
\alias{addYears}
\title{Add years to series}
\usage{
addYears(lastYear, densProfile)
}
\arguments{
\item{lastYear}{Last year of series}

\item{densProfile}{Density profile}
}
\value{
Density profile with years
}
\description{
Add years to series
}
