% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvgDens.R
\name{calcAvgDens}
\alias{calcAvgDens}
\title{Calculate average wood, earlywood and latewood density for every ring}
\usage{
calcAvgDens(densProfile)
}
\arguments{
\item{densProfile}{Density profile}
}
\value{
Density profile with average values added
}
\description{
Calculate average wood, earlywood and latewood density for every ring
}
