% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProfile.R
\name{checkProfile}
\alias{checkProfile}
\title{Verify position of ring transitions of a density profile}
\usage{
checkProfile(profile_with_borders, totRings)
}
\arguments{
\item{profile_with_borders}{xRing profile with transitions between rings located}

\item{totRings}{Total number of rings of the disk}
}
\value{
xRing profile with corrected ring location
}
\description{
Verify position of ring transitions of a density profile
}
\examples{
t2 <- checkProfile(t1, 30)

}
