% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteRingFromImage.R
\name{deleteRingFromImage}
\alias{deleteRingFromImage}
\title{Add ring to pith to bark profile from CT scan image}
\usage{
deleteRingFromImage(n = 1, densProfile, im)
}
\arguments{
\item{n}{Number of rings to remove}

\item{densProfile}{Density profile}

\item{im}{Density matrix}
}
\value{
Corrected density profile with ring(s) removed and red bar in plot of deleted ring
}
\description{
Add ring to pith to bark profile from CT scan image
}
