% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteRingFromProfile.R
\name{deleteRingFromProfile}
\alias{deleteRingFromProfile}
\title{Delete ring from a pith to bark profile}
\usage{
deleteRingFromProfile(n = 1, densProfile)
}
\arguments{
\item{n}{Number of rings to remove}

\item{densProfile}{Density profile}
}
\value{
Corrected density profile with ring(s) removed and red bar in plot of deleted ring
}
\description{
Delete ring from a pith to bark profile
}
