% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityDataFrame.R
\name{densityDataFrame}
\alias{densityDataFrame}
\title{Convert to dataframe}
\usage{
densityDataFrame(densProfile, sampleID = "NoID")
}
\arguments{
\item{densProfile}{Density profile}

\item{sampleID}{Sample ID}
}
\value{
Dataframe with cambial age, density, years
}
\description{
Convert to dataframe
}
