% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensityProfile.R
\name{getDensityProfile}
\alias{getDensityProfile}
\title{Obtain pith to bark density profile}
\usage{
getDensityProfile(
  pith_coord_units,
  im_raster,
  theta.deg = 0,
  toPlot = FALSE,
  density_cutoff = 0.1,
  name = "t1"
)
}
\arguments{
\item{pith_coord_units}{Coordinates of the pith in vector (x, y)}

\item{im_raster}{Raster image of the CT scan image}

\item{theta.deg}{Angle of the path (0-360 deg)}

\item{toPlot}{Boolean to determine if a plot of the path on the image is returned}

\item{density_cutoff}{Density value below which values are not considered}
}
\value{
Dataframe with profile values (distance from pith, density)
}
\description{
Obtain pith to bark density profile
}
\examples{
path_profile <- getDensityProfile(theta = 0,
  pith_coord_units = pith_coord_units,
  im_raster = im_raster,
  toPlot = TRUE)

}
