% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grayToDensity.R
\name{grayToDensity}
\alias{grayToDensity}
\title{Convert from 8bit gray scale to density}
\usage{
grayToDensity(im, a = -0.1321, b = 0.01834)
}
\arguments{
\item{im}{Matrix of CT scan image in 8bit gray scale}

\item{a}{Intercept of the calibration curve}

\item{b}{Slope of the calibration curve}
}
\value{
Matrix of density values
}
\description{
Convert from 8bit gray scale to density
}
