% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageToMatrix.R
\name{imageToMatrix}
\alias{imageToMatrix}
\title{Convert dicom image to matrix}
\usage{
imageToMatrix(img)
}
\arguments{
\item{img}{Dicom image}
}
\value{
Matrix of image
}
\description{
Convert dicom image to matrix
}
