% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatePathEnd.R
\name{locatePathEnd}
\alias{locatePathEnd}
\title{Get coordinates of the end of the path on a CT scan image}
\usage{
locatePathEnd(im, pithCoord)
}
\arguments{
\item{im}{CT scan image}
}
\value{
Coordinates of the end of the path
}
\description{
Get coordinates of the end of the path on a CT scan image
}
