% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixToRaster.R
\name{matrixToRaster}
\alias{matrixToRaster}
\title{Convert image matrix to raster}
\usage{
matrixToRaster(im, pixel_size_x, pixel_size_y)
}
\arguments{
\item{im}{Density matrix}

\item{pixel_size_x}{Pith coordinate in x in unit lengths}

\item{pixel_size_y}{Pith coordinates in y in unit lengths}
}
\value{
Density raster
}
\description{
Convert image matrix to raster
}
