% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImageProfile.R
\name{plotImageProfile}
\alias{plotImageProfile}
\title{Plot scan image, profile path and ring limits}
\usage{
plotImageProfile(densProfile, im)
}
\arguments{
\item{densProfile}{Density profile}

\item{im}{Density matrix}
}
\value{
Plot of ring limits on scan image
}
\description{
Plot scan image, profile path and ring limits
}
