% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relToPixel.R
\name{relToPixel}
\alias{relToPixel}
\title{Change from relative to fixed pixel coordinate system}
\usage{
relToPixel(pith_coord, im)
}
\arguments{
\item{pith_coord}{Pith coordinates in relative space (x, y)}

\item{im}{Density matrix}
}
\value{
Pixel coordinates in number of pixels (x, y)
}
\description{
Change from relative to fixed pixel coordinate system
}
