% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLastYear.R
\name{removeLastYear}
\alias{removeLastYear}
\title{Remove the last year of a profile}
\usage{
removeLastYear(densProfile)
}
\arguments{
\item{densProfile}{Density profile}
}
\value{
Density profile with the last year removed
}
\description{
Remove the last year of a profile
}
