% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringBorders.R
\name{ringBorders}
\alias{ringBorders}
\title{Wrapper to use getBorders of xRing package to identify transition points between rings}
\usage{
ringBorders(t, name = "t1", thershold = 0.1, minTrw = 1)
}
\arguments{
\item{t}{Dataframe with density profiles}

\item{name}{Name of the series}

\item{thershold}{Thershold used by getBorders to identify transistion between rings}

\item{minTrw}{Width of the narrowest ring in number of measurements}
}
\value{
xRing list for the profile
}
\description{
Wrapper to use getBorders of xRing package to identify transition points between rings
}
