% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyPith.R
\name{verifyPith}
\alias{verifyPith}
\title{Check if pith location is correct}
\usage{
verifyPith(im, pith_coord)
}
\arguments{
\item{im}{Density matrix of image}

\item{pith_coord}{Pith coordinates}
}
\value{
Corrected pith coordinates
}
\description{
Check if pith location is correct
}
