% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBitTo8Bit.R
\name{xBitTo8Bit}
\alias{xBitTo8Bit}
\title{Convert gray scale from measured bits to 8bit}
\usage{
xBitTo8Bit(im, bits)
}
\arguments{
\item{im}{Matrix of values in x bits}

\item{bits}{Number of bits of the original gray scale}
}
\value{
Matrix of gray scale values in 8bits
}
\description{
Convert gray scale from measured bits to 8bit
}
