% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractProfile.R
\name{extractProfile}
\alias{extractProfile}
\title{Get profile between two points of the CTScan image matrix}
\usage{
extractProfile(
  im,
  imHeader,
  beginPath,
  endPath,
  r = 10,
  k = 2,
  threshold = 0.01
)
}
\arguments{
\item{im}{Density matrix}

\item{r}{Profile width}

\item{k}{Rolling window width, integer}

\item{threshold}{Threshold value between maximum and minimum density to establish change of ring}

\item{x}{Values of X for points 1 and 2 (x1, x2)}

\item{y}{Values of y for points 1 and 2 (y1, y2)}
}
\description{
Get profile between two points of the CTScan image matrix
}
