% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEwLw.R
\name{getEwLw}
\alias{getEwLw}
\title{Establish the transition point from earlywood to latewood for a series of rings}
\usage{
getEwLw(xRingList, dist)
}
\arguments{
\item{xRingList}{A list of xRing profiles}

\item{dist}{A vector of distances of each point from pith center}
}
\value{
xRingList with EW to LW transition points
}
\description{
Establish the transition point from earlywood to latewood for a series of rings
}
