% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pithCoordinates.R
\name{pithCoordinates}
\alias{pithCoordinates}
\title{convert pith coordinates from pixels to length units}
\usage{
pithCoordinates(pith_coord, pixel_size_x, pixel_size_y)
}
\arguments{
\item{pith_coord}{Pith coordinates in pixels}

\item{pixel_size_x}{Pixel size in x}

\item{pixel_size_y}{Pixel size in y}
}
\value{
Pixel coordinates in length units
}
\description{
convert pith coordinates from pixels to length units
}
